refactor(router): consolidate SSR string renderers into single renderToString
Keep the cached-platform implementation (formerly renderToStringFast)
as the one renderToString export. Drop the per-request renderApplication
path and the V1 token renderer; the V2 renderer becomes string-renderer.ts
with the V2 suffixes removed.
