+++ packages/router/server/src/index.ts
@@ -5,11 +5,7 @@ export {
   renderServerComponent,
 } from './server-component-render';
 export { render } from './render';
-export { renderToString } from './render-to-string';
-export {
-  renderToStringFast,
-  destroySharedPlatform,
-} from './render-to-string-fast';
+export { renderToString, destroySharedPlatform } from './render-to-string';
 
 // Server Functions (issue #2422) — server authoring + dispatch runtime.
 export { serverFn } from './server-fn/server-fn';
